PayGo/web-service/backend/requirements.txt
# Основной веб-фреймворк
fastapi==0.100.0
uvicorn[standard]==0.23.1
starlette==0.27.0

# База данных и ORM
sqlalchemy==2.0.18
alembic==1.11.1
psycopg2-binary==2.9.6
databases[postgresql]==0.7.0

# Аутентификация и безопасность
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1

# API и интеграция с банками
httpx==0.24.1
requests==2.31.0
aiohttp==3.8.4

# Валидация данных
pydantic[email]==2.0.3
pydantic-settings==2.0.2

# CORS и middleware
python-cors==1.7.0

# Работа с JSON и файлами
orjson==3.9.2
python-dotenv==1.0.0

# Биометрия (серверная часть)
opencv-python==4.8.1.78
numpy==1.24.3

# Мониторинг и логирование
loguru==0.7.0
prometheus-client==0.17.1

# Тестирование
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.24.1

# Развертывание
gunicorn==21.2.0
docker==6.1.3

# Работа с изображениями
Pillow==10.0.0

# Кэширование
redis==4.6.0
hiredis==2.2.3 